Fixed Sprites
Still no sprite for the changing world platforms.
Still no sprite for the pressure plate.
Still no sprite for the doors that move up. (Unless it's that metal door one, but still, it's barely noticeable.)
Level 1 is too easy, broken in some spots, and messy. (The player can jump over the key door and get to the exit without trouble.)
Level 2 is too easy and too basic.
The wrong "Both World Platforms" were used in the levels, so the sprites are still messed up.
There are no sfx in the game at all. (I will try to get them in before 9.)
+ Fixed a few sprites that were never fixed before.